+++ elevation_test.go
@@ -17,11 +17,13 @@ func TestNewElevationControl(t *testing.T) {
 	originalIsElevated := cmdtoolkit.IsElevated
 	originalIsTerminal := cmdtoolkit.IsTerminal
 	originalIsCygwinTerminal := cmdtoolkit.IsCygwinTerminal
+	originalReadAlternateDataStream := cmdtoolkit.ReadAlternateDataStream
 	defer func() {
 		cmdtoolkit.GetCurrentProcessToken = originalGetCurrentProcessToken
 		cmdtoolkit.IsElevated = originalIsElevated
 		cmdtoolkit.IsTerminal = originalIsTerminal
 		cmdtoolkit.IsCygwinTerminal = originalIsCygwinTerminal
+		cmdtoolkit.ReadAlternateDataStream = originalReadAlternateDataStream
 	}()
 	cmdtoolkit.GetCurrentProcessToken = func() (t windows.Token) {
 		return
@@ -29,12 +31,14 @@ func TestNewElevationControl(t *testing.T) {
 	tests := map[string]struct {
 		assertElevated   bool
 		assertRedirected bool
+		assertForbidden  bool
 		wantStatus       []string
 		output.WantedRecording
 	}{
-		"neither elevated nor redirected": {
+		"neither elevated nor redirected, nor forbidden": {
 			assertElevated:   false,
 			assertRedirected: false,
+			assertForbidden:  false,
 			wantStatus: []string{
 				"myApp is not running with elevated privileges",
 			},
@@ -42,6 +46,10 @@ func TestNewElevationControl(t *testing.T) {
 				Log: "" +
 					"level='info'" +
 					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='false'" +
+					" ads_id=''" +
 					" elevated='false'" +
 					" environment_variable=''" +
 					" stderr_redirected='false'" +
@@ -50,9 +58,10 @@ func TestNewElevationControl(t *testing.T) {
 					" msg='elevation state'\n",
 			},
 		},
-		"not elevated, but redirected": {
+		"not elevated, but redirected, and not forbidden": {
 			assertElevated:   false,
 			assertRedirected: true,
+			assertForbidden:  false,
 			wantStatus: []string{
 				"myApp is not running with elevated privileges",
 				"stderr, stdin, and stdout have been redirected",
@@ -61,6 +70,10 @@ func TestNewElevationControl(t *testing.T) {
 				Log: "" +
 					"level='info'" +
 					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='false'" +
+					" ads_id=''" +
 					" elevated='false'" +
 					" environment_variable=''" +
 					" stderr_redirected='true'" +
@@ -69,9 +82,10 @@ func TestNewElevationControl(t *testing.T) {
 					" msg='elevation state'\n",
 			},
 		},
-		"elevated, but not redirected": {
+		"elevated, but not redirected, and not forbidden": {
 			assertElevated:   true,
 			assertRedirected: false,
+			assertForbidden:  false,
 			wantStatus: []string{
 				"myApp is running with elevated privileges",
 			},
@@ -79,6 +93,10 @@ func TestNewElevationControl(t *testing.T) {
 				Log: "" +
 					"level='info'" +
 					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='false'" +
+					" ads_id=''" +
 					" elevated='true'" +
 					" environment_variable=''" +
 					" stderr_redirected='false'" +
@@ -87,9 +105,10 @@ func TestNewElevationControl(t *testing.T) {
 					" msg='elevation state'\n",
 			},
 		},
-		"elevated and redirected": {
+		"elevated and redirected and not forbidden": {
 			assertElevated:   true,
 			assertRedirected: true,
+			assertForbidden:  false,
 			wantStatus: []string{
 				"myApp is running with elevated privileges",
 			},
@@ -97,6 +116,105 @@ func TestNewElevationControl(t *testing.T) {
 				Log: "" +
 					"level='info'" +
 					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='false'" +
+					" ads_id=''" +
+					" elevated='true'" +
+					" environment_variable=''" +
+					" stderr_redirected='true'" +
+					" stdin_redirected='true'" +
+					" stdout_redirected='true'" +
+					" msg='elevation state'\n",
+			},
+		},
+		"neither elevated nor redirected but forbidden": {
+			assertElevated:   false,
+			assertRedirected: false,
+			assertForbidden:  true,
+			wantStatus: []string{
+				"myApp is not running with elevated privileges",
+				"The zone id (Internet) forbids myApp from running with elevated privileges",
+			},
+			WantedRecording: output.WantedRecording{
+				Log: "" +
+					"level='info'" +
+					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='true'" +
+					" ads_id='Internet'" +
+					" elevated='false'" +
+					" environment_variable=''" +
+					" stderr_redirected='false'" +
+					" stdin_redirected='false'" +
+					" stdout_redirected='false'" +
+					" msg='elevation state'\n",
+			},
+		},
+		"not elevated, but redirected and forbidden": {
+			assertElevated:   false,
+			assertRedirected: true,
+			assertForbidden:  true,
+			wantStatus: []string{
+				"myApp is not running with elevated privileges",
+				"stderr, stdin, and stdout have been redirected",
+				"The zone id (Internet) forbids myApp from running with elevated privileges",
+			},
+			WantedRecording: output.WantedRecording{
+				Log: "" +
+					"level='info'" +
+					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='true'" +
+					" ads_id='Internet'" +
+					" elevated='false'" +
+					" environment_variable=''" +
+					" stderr_redirected='true'" +
+					" stdin_redirected='true'" +
+					" stdout_redirected='true'" +
+					" msg='elevation state'\n",
+			},
+		},
+		"elevated, but not redirected but forbidden": {
+			assertElevated:   true,
+			assertRedirected: false,
+			assertForbidden:  true,
+			wantStatus: []string{
+				"myApp is running with elevated privileges",
+			},
+			WantedRecording: output.WantedRecording{
+				Log: "" +
+					"level='info'" +
+					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='true'" +
+					" ads_id='Internet'" +
+					" elevated='true'" +
+					" environment_variable=''" +
+					" stderr_redirected='false'" +
+					" stdin_redirected='false'" +
+					" stdout_redirected='false'" +
+					" msg='elevation state'\n",
+			},
+		},
+		"elevated and redirected but forbidden": {
+			assertElevated:   true,
+			assertRedirected: true,
+			assertForbidden:  true,
+			wantStatus: []string{
+				"myApp is running with elevated privileges",
+			},
+			WantedRecording: output.WantedRecording{
+				Log: "" +
+					"level='info'" +
+					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='true'" +
+					" ads_id='Internet'" +
 					" elevated='true'" +
 					" environment_variable=''" +
 					" stderr_redirected='true'" +
@@ -111,6 +229,20 @@ func TestNewElevationControl(t *testing.T) {
 			cmdtoolkit.IsElevated = func(_ windows.Token) bool { return tt.assertElevated }
 			cmdtoolkit.IsTerminal = func(_ uintptr) bool { return !tt.assertRedirected }
 			cmdtoolkit.IsCygwinTerminal = cmdtoolkit.IsTerminal
+			if tt.assertForbidden {
+				cmdtoolkit.ReadAlternateDataStream = func() *cmdtoolkit.ADSInformation {
+					return &cmdtoolkit.ADSInformation{
+						Forbidden: tt.assertForbidden,
+						ID:        "Internet",
+					}
+				}
+			} else {
+				cmdtoolkit.ReadAlternateDataStream = func() *cmdtoolkit.ADSInformation {
+					return &cmdtoolkit.ADSInformation{
+						Forbidden: tt.assertForbidden,
+					}
+				}
+			}
 			ec := cmdtoolkit.NewElevationControl()
 			o := output.NewRecorder()
 			ec.Log(o, output.Info)
@@ -129,12 +261,14 @@ func TestNewElevationControlWithEnvVar(t *testing.T) {
 	originalIsElevated := cmdtoolkit.IsElevated
 	originalIsTerminal := cmdtoolkit.IsTerminal
 	originalIsCygwinTerminal := cmdtoolkit.IsCygwinTerminal
+	originalReadAlternateDataStream := cmdtoolkit.ReadAlternateDataStream
 	defer func() {
 		envVarMemento.Restore()
 		cmdtoolkit.GetCurrentProcessToken = originalGetCurrentProcessToken
 		cmdtoolkit.IsElevated = originalIsElevated
 		cmdtoolkit.IsTerminal = originalIsTerminal
 		cmdtoolkit.IsCygwinTerminal = originalIsCygwinTerminal
+		cmdtoolkit.ReadAlternateDataStream = originalReadAlternateDataStream
 	}()
 	_ = os.Unsetenv(envVarName)
 	cmdtoolkit.GetCurrentProcessToken = func() (t windows.Token) {
@@ -144,9 +278,211 @@ func TestNewElevationControlWithEnvVar(t *testing.T) {
 		defaultEnvVarValue bool
 		assertElevated     bool
 		assertRedirected   bool
+		assertForbidden    bool
 		wantStatus         []string
 		output.WantedRecording
 	}{
+		"env var false, neither elevated nor redirected, forbidden": {
+			defaultEnvVarValue: false,
+			assertElevated:     false,
+			assertRedirected:   false,
+			assertForbidden:    true,
+			wantStatus: []string{
+				"myApp is not running with elevated privileges",
+				"The environment variable MYAPP_RUNS_AS_ADMIN evaluates as false",
+				"The zone id (Restricted sites) forbids myApp from running with elevated privileges",
+			},
+			WantedRecording: output.WantedRecording{
+				Log: "" +
+					"level='info'" +
+					" admin_permission='false'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='true'" +
+					" ads_id='Restricted sites'" +
+					" elevated='false'" +
+					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
+					" stderr_redirected='false'" +
+					" stdin_redirected='false'" +
+					" stdout_redirected='false'" +
+					" msg='elevation state'\n",
+			},
+		},
+		"env var false, not elevated, but redirected, forbidden": {
+			defaultEnvVarValue: false,
+			assertElevated:     false,
+			assertRedirected:   true,
+			assertForbidden:    true,
+			wantStatus: []string{
+				"myApp is not running with elevated privileges",
+				"stderr, stdin, and stdout have been redirected",
+				"The environment variable MYAPP_RUNS_AS_ADMIN evaluates as false",
+				"The zone id (Restricted sites) forbids myApp from running with elevated privileges",
+			},
+			WantedRecording: output.WantedRecording{
+				Log: "" +
+					"level='info'" +
+					" admin_permission='false'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='true'" +
+					" ads_id='Restricted sites'" +
+					" elevated='false'" +
+					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
+					" stderr_redirected='true'" +
+					" stdin_redirected='true'" +
+					" stdout_redirected='true'" +
+					" msg='elevation state'\n",
+			},
+		},
+		"env var false, elevated, but not redirected, forbidden": {
+			defaultEnvVarValue: false,
+			assertElevated:     true,
+			assertRedirected:   false,
+			assertForbidden:    true,
+			wantStatus: []string{
+				"myApp is running with elevated privileges",
+			},
+			WantedRecording: output.WantedRecording{
+				Log: "" +
+					"level='info'" +
+					" admin_permission='false'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='true'" +
+					" ads_id='Restricted sites'" +
+					" elevated='true'" +
+					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
+					" stderr_redirected='false'" +
+					" stdin_redirected='false'" +
+					" stdout_redirected='false'" +
+					" msg='elevation state'\n",
+			},
+		},
+		"env var false, elevated and redirected, forbidden": {
+			defaultEnvVarValue: false,
+			assertElevated:     true,
+			assertRedirected:   true,
+			assertForbidden:    true,
+			wantStatus: []string{
+				"myApp is running with elevated privileges",
+			},
+			WantedRecording: output.WantedRecording{
+				Log: "" +
+					"level='info'" +
+					" admin_permission='false'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='true'" +
+					" ads_id='Restricted sites'" +
+					" elevated='true'" +
+					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
+					" stderr_redirected='true'" +
+					" stdin_redirected='true'" +
+					" stdout_redirected='true'" +
+					" msg='elevation state'\n",
+			},
+		},
+		"env var true, neither elevated nor redirected, forbidden": {
+			defaultEnvVarValue: true,
+			assertElevated:     false,
+			assertRedirected:   false,
+			assertForbidden:    true,
+			wantStatus: []string{
+				"myApp is not running with elevated privileges",
+				"The zone id (Restricted sites) forbids myApp from running with elevated privileges",
+			},
+			WantedRecording: output.WantedRecording{
+				Log: "" +
+					"level='info'" +
+					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='true'" +
+					" ads_id='Restricted sites'" +
+					" elevated='false'" +
+					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
+					" stderr_redirected='false'" +
+					" stdin_redirected='false'" +
+					" stdout_redirected='false'" +
+					" msg='elevation state'\n",
+			},
+		},
+		"env var true, not elevated, but redirected, forbidden": {
+			defaultEnvVarValue: true,
+			assertElevated:     false,
+			assertRedirected:   true,
+			assertForbidden:    true,
+			wantStatus: []string{
+				"myApp is not running with elevated privileges",
+				"stderr, stdin, and stdout have been redirected",
+				"The zone id (Restricted sites) forbids myApp from running with elevated privileges",
+			},
+			WantedRecording: output.WantedRecording{
+				Log: "" +
+					"level='info'" +
+					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='true'" +
+					" ads_id='Restricted sites'" +
+					" elevated='false'" +
+					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
+					" stderr_redirected='true'" +
+					" stdin_redirected='true'" +
+					" stdout_redirected='true'" +
+					" msg='elevation state'\n",
+			},
+		},
+		"env var true, elevated, but not redirected, forbidden": {
+			defaultEnvVarValue: true,
+			assertElevated:     true,
+			assertRedirected:   false,
+			assertForbidden:    true,
+			wantStatus: []string{
+				"myApp is running with elevated privileges",
+			},
+			WantedRecording: output.WantedRecording{
+				Log: "" +
+					"level='info'" +
+					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='true'" +
+					" ads_id='Restricted sites'" +
+					" elevated='true'" +
+					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
+					" stderr_redirected='false'" +
+					" stdin_redirected='false'" +
+					" stdout_redirected='false'" +
+					" msg='elevation state'\n",
+			},
+		},
+		"env var true, elevated and redirected, forbidden": {
+			defaultEnvVarValue: true,
+			assertElevated:     true,
+			assertRedirected:   true,
+			assertForbidden:    true,
+			wantStatus: []string{
+				"myApp is running with elevated privileges",
+			},
+			WantedRecording: output.WantedRecording{
+				Log: "" +
+					"level='info'" +
+					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='true'" +
+					" ads_id='Restricted sites'" +
+					" elevated='true'" +
+					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
+					" stderr_redirected='true'" +
+					" stdin_redirected='true'" +
+					" stdout_redirected='true'" +
+					" msg='elevation state'\n",
+			},
+		},
+
 		"env var false, neither elevated nor redirected": {
 			defaultEnvVarValue: false,
 			assertElevated:     false,
@@ -159,6 +495,10 @@ func TestNewElevationControlWithEnvVar(t *testing.T) {
 				Log: "" +
 					"level='info'" +
 					" admin_permission='false'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='false'" +
+					" ads_id=''" +
 					" elevated='false'" +
 					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
 					" stderr_redirected='false'" +
@@ -180,6 +520,10 @@ func TestNewElevationControlWithEnvVar(t *testing.T) {
 				Log: "" +
 					"level='info'" +
 					" admin_permission='false'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='false'" +
+					" ads_id=''" +
 					" elevated='false'" +
 					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
 					" stderr_redirected='true'" +
@@ -199,6 +543,10 @@ func TestNewElevationControlWithEnvVar(t *testing.T) {
 				Log: "" +
 					"level='info'" +
 					" admin_permission='false'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='false'" +
+					" ads_id=''" +
 					" elevated='true'" +
 					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
 					" stderr_redirected='false'" +
@@ -218,6 +566,10 @@ func TestNewElevationControlWithEnvVar(t *testing.T) {
 				Log: "" +
 					"level='info'" +
 					" admin_permission='false'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='false'" +
+					" ads_id=''" +
 					" elevated='true'" +
 					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
 					" stderr_redirected='true'" +
@@ -237,6 +589,10 @@ func TestNewElevationControlWithEnvVar(t *testing.T) {
 				Log: "" +
 					"level='info'" +
 					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='false'" +
+					" ads_id=''" +
 					" elevated='false'" +
 					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
 					" stderr_redirected='false'" +
@@ -257,6 +613,10 @@ func TestNewElevationControlWithEnvVar(t *testing.T) {
 				Log: "" +
 					"level='info'" +
 					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='false'" +
+					" ads_id=''" +
 					" elevated='false'" +
 					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
 					" stderr_redirected='true'" +
@@ -276,6 +636,10 @@ func TestNewElevationControlWithEnvVar(t *testing.T) {
 				Log: "" +
 					"level='info'" +
 					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='false'" +
+					" ads_id=''" +
 					" elevated='true'" +
 					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
 					" stderr_redirected='false'" +
@@ -295,6 +659,10 @@ func TestNewElevationControlWithEnvVar(t *testing.T) {
 				Log: "" +
 					"level='info'" +
 					" admin_permission='true'" +
+					" ads_content=''" +
+					" ads_error='<nil>'" +
+					" ads_forbidden='false'" +
+					" ads_id=''" +
 					" elevated='true'" +
 					" environment_variable='MYAPP_RUNS_AS_ADMIN'" +
 					" stderr_redirected='true'" +
@@ -309,6 +677,20 @@ func TestNewElevationControlWithEnvVar(t *testing.T) {
 			cmdtoolkit.IsElevated = func(_ windows.Token) bool { return tt.assertElevated }
 			cmdtoolkit.IsTerminal = func(_ uintptr) bool { return !tt.assertRedirected }
 			cmdtoolkit.IsCygwinTerminal = cmdtoolkit.IsTerminal
+			if tt.assertForbidden {
+				cmdtoolkit.ReadAlternateDataStream = func() *cmdtoolkit.ADSInformation {
+					return &cmdtoolkit.ADSInformation{
+						Forbidden: true,
+						ID:        "Restricted sites",
+					}
+				}
+			} else {
+				cmdtoolkit.ReadAlternateDataStream = func() *cmdtoolkit.ADSInformation {
+					return &cmdtoolkit.ADSInformation{
+						Forbidden: false,
+					}
+				}
+			}
 			o := output.NewRecorder()
 			ec := cmdtoolkit.NewElevationControlWithEnvVar(envVarName, tt.defaultEnvVarValue)
 			ec.Log(o, output.Info)